Went in and changed any double to BigDecimal.

+++ src/main/java/com/example/project_1/dto/RevenueDTO.java
@@ -2,9 +2,11 @@ package com.example.project_1.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RevenueDTO {
     private Long event_Id;
     private String event_Name;
-    private Double total_Revenue;
+    private BigDecimal total_Revenue;
 }
